<feat> реализовать получение списка заявок на свое событие + рефактор дубликатов кода

+++ ewm-main-service/src/main/java/ru/practicum/dto/event/NewEventDto.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 @Getter
 public class NewEventDto {
     private String annotation;
-    private Integer category;
+    private Long category;
     private String description;
     private LocalDateTime eventDate;
     private Location location;
